Update readmes with new backend plugins system
