+++ plugins/graphql-backend/api-report.md
@@ -3,50 +3,55 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
 import { BatchLoadFn } from '@backstage/plugin-graphql-common';
-import { Config } from '@backstage/config';
-import express from 'express';
+import { ExtensionPoint } from '@backstage/backend-plugin-api';
 import { GraphQLContext } from '@backstage/plugin-graphql-common';
-import { Logger } from 'winston';
 import { Module } from 'graphql-modules';
 import { Options } from 'dataloader';
 import { Plugin as Plugin_2 } from 'graphql-yoga';
 
 // @public (undocumented)
-export function createRouter<TContext extends Record<string, any>>({
-  config,
-  logger,
-  schema,
-  modules,
-  plugins,
-  loaders,
-  dataloaderOptions,
-  additionalContext,
-  generateOpaqueTypes,
-}: RouterOptions<TContext>): Promise<express.Router>;
+export interface GraphQLApplicationExtensionPoint {
+  // (undocumented)
+  addModule(
+    module: (() => Module | Promise<Module>) | Module | Promise<Module>,
+  ): void;
+  // (undocumented)
+  addSchema(schema: string): void;
+}
 
 // @public (undocumented)
-export interface RouterOptions<TContext extends Record<string, any>> {
+export const graphqlApplicationExtensionPoint: ExtensionPoint<GraphQLApplicationExtensionPoint>;
+
+// @public (undocumented)
+export const graphqlPlugin: (
+  options?: RouterOptions | undefined,
+) => BackendFeature;
+
+// @public (undocumented)
+export interface GraphQLYogaExtensionPoint {
   // (undocumented)
-  additionalContext?:
-    | ((initialContext: GraphQLContext) => TContext | Promise<TContext>)
-    | Promise<TContext>
-    | TContext;
+  addLoader(name: string, loader: BatchLoadFn<GraphQLContext>): void;
   // (undocumented)
-  config: Config;
+  addPlugin(plugin: Plugin_2): void;
   // (undocumented)
-  dataloaderOptions?: Options<string, any>;
+  setContext<TContext extends Record<string, any>>(
+    context:
+      | ((initialContext: GraphQLContext) => TContext | Promise<TContext>)
+      | Promise<TContext>
+      | TContext,
+  ): void;
+  // (undocumented)
+  setDataloaderOptions(options: Options<string, any>): void;
+}
+
+// @public (undocumented)
+export const graphqlYogaExtensionPoint: ExtensionPoint<GraphQLYogaExtensionPoint>;
+
+// @public (undocumented)
+export interface RouterOptions {
   // (undocumented)
   generateOpaqueTypes?: boolean;
-  // (undocumented)
-  loaders?: Record<string, BatchLoadFn<TContext & GraphQLContext>>;
-  // (undocumented)
-  logger: Logger;
-  // (undocumented)
-  modules?: ((() => Module | Promise<Module>) | Module | Promise<Module>)[];
-  // (undocumented)
-  plugins?: Plugin_2[];
-  // (undocumented)
-  schema?: string | string[];
 }
 ```
