+++ plugins/graphql-backend/src/extensions/graphqlApplicationExtension.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { createExtensionPoint } from '@backstage/backend-plugin-api';
+import { Module } from 'graphql-modules';
 
-/**
- * An experimental Backstage backend plugin for GraphQL
- *
- * @packageDocumentation
- */
+/** @public */
+export interface GraphQLApplicationExtensionPoint {
+  addSchema(schema: string): void;
+  addModule(
+    module: (() => Module | Promise<Module>) | Module | Promise<Module>,
+  ): void;
+}
 
-export * from './graphql';
-export * from './extensions';
+/** @public */
+export const graphqlApplicationExtensionPoint =
+  createExtensionPoint<GraphQLApplicationExtensionPoint>({
+    id: 'graphql.application',
+  });
